Revert "New try page"
