+++ editor/elm.json
@@ -11,10 +11,7 @@
             "elm/html": "1.0.0",
             "elm/http": "2.0.0",
             "elm/json": "1.1.3",
-            "elm/parser": "1.1.0",
-            "elm/project-metadata-utils": "1.0.2",
-            "elm/svg": "1.0.1",
-            "feathericons/elm-feather": "1.5.0"
+            "elm/parser": "1.1.0"
         },
         "indirect": {
             "elm/bytes": "1.0.8",
