--- editor/src/Data/Problem.elm
@@ -1,175 +1,0 @@
-module Data.Problem exposing (..)
-
-{-| The formatting of compilation errors.
-
--}
-
-import Html exposing (..)
-import Html.Attributes exposing (..)
-import Html.Events exposing (onClick, on)
-import Elm.Error as Error
-import FeatherIcons as I
-
-
-
--- PROCESSING
-
-
-type alias Problem =
-  { index : Int
-  , location : Location
-  , title : String
-  , message : List Error.Chunk
-  }
-
-
-type Location
-  = General { path : Maybe String }
-  | Specific { path : String, name : String, region : Error.Region }
-
-
-toIndexedProblems : Error.Error -> List  Problem
-toIndexedProblems errors =
-  case errors of
-    Error.GeneralProblem problem ->
-      [ { index = 0
-        , location = General { path = problem.path }
-        , title = problem.title
-        , message = problem.message
-        }
-      ]
-
-    Error.ModuleProblems modules ->
-      let toModuleProblems module_ ( nextIndex, all ) =
-            List.foldr (toSingleProblem module_) ( nextIndex, all ) module_.problems
-
-          toSingleProblem module_ problem ( nextIndex, all ) =
-            let indexedProblem =
-                  { index = nextIndex
-                  , location =
-                      Specific
-                        { path = module_.path
-                        , name = module_.name
-                        , region = problem.region
-                        }
-                  , title = problem.title
-                  , message = problem.message
-                  }
-            in
-            ( nextIndex + 1, indexedProblem :: all )
-      in
-      List.foldr toModuleProblems ( 0, [] ) modules
-        |> Tuple.second
-
-
-getRegion : Problem -> Maybe Error.Region
-getRegion problem =
-  case problem.location of
-    General _ -> Nothing
-    Specific specific -> Just specific.region
-
-
-
--- MANY
-
-
-type alias Problems =
-  ( List Problem, Problem, List Problem )
-
-
-init : List Problem -> Maybe Problems
-init pbs =
-  case pbs of
-    first :: rest -> Just ( [], first, rest )
-    _ -> Nothing
-
-
-getAll : Problems -> List Problem
-getAll ( prev, curr, next ) =
-  List.reverse prev ++ [ curr ] ++ next
-
-
-getFocused : Problems -> Problem
-getFocused ( _, current, _ ) =
-  current
-
-
-getPrevious : Problems -> Maybe Problem
-getPrevious ( prev, _, _ ) =
-  case prev of
-    first :: _ -> Just first
-    [] -> Nothing
-
-
-getNext : Problems -> Maybe Problem
-getNext ( _, _, next ) =
-  case next of
-    first :: _ -> Just first
-    [] -> Nothing
-
-
-hasPrevious : Problems -> Bool
-hasPrevious ( prev, curr, next ) =
-  not (List.isEmpty prev)
-
-
-hasNext : Problems -> Bool
-hasNext ( prev, curr, next ) =
-  not (List.isEmpty next)
-
-
-focusNext : Problems -> Problems
-focusNext pbs =
-  case pbs of
-    ( prev, curr, next :: remaining ) -> ( curr :: prev, next, remaining )
-    ( prev, curr, [] ) -> pbs
-
-
-focusPrevious : Problems -> Problems
-focusPrevious pbs =
-  case pbs of
-    ( prev :: remaining, curr, next ) -> ( remaining, prev, curr :: next )
-    ( [], curr, next ) -> pbs
-
-
-
--- HELPERS
-
-
-getSummary : Problem -> String
-getSummary problem =
-  let plainText = getPlainText problem
-      firstLine = List.head (String.split "\n\n" plainText)
-      firstColon = List.head (String.split ":" plainText)
-  in
-  case ( firstLine, firstColon ) of
-    ( Just firstLine_, Just firstColon_ ) ->
-      if String.length firstLine_ > String.length firstColon_ then firstColon_ else firstLine_
-
-    ( Just firstLine_, Nothing ) ->
-      firstLine_
-
-    ( Nothing, Just firstColon_ ) ->
-      firstColon_
-
-    ( Nothing, Nothing ) ->
-      "Click to see full error message!"
-
-
-getPlainText : Problem -> String
-getPlainText problem =
-  let toPlain chunks =
-        case chunks of
-          [] ->
-            ""
-
-          chunk :: rest ->
-            case chunk of
-              Error.Unstyled string ->
-                string ++ toPlain rest
-
-              Error.Styled _ string ->
-                string ++ toPlain rest
-  in
-  toPlain problem.message
-
