--- editor/src/Data/Window.elm
@@ -1,7 +1,0 @@
-module Data.Window exposing (Window)
-
-
-type alias Window =
-  { width : Int
-  , height : Int
-  }
